Extract shared logic of edit form to separate component
