+++ frontend/src/pages/application/PipelineStage/create.vue
@@ -15,88 +15,34 @@
         </SideNavigation>
     </Teleport>
     <main>
-        <ff-loading
-            v-if="loading"
-            message="Creating Pipeline Stage..."
+        <PipelineStageForm
+            :application="application"
+            :pipeline="pipeline"
+            :stage="{}"
+            @submit="create"
         />
-        <form
-            v-else
-            class="space-y-6"
-            @submit.prevent="create"
-        >
-            <SectionTopMenu
-                :hero="'Add Pipeline Stage'"
-            />
-
-            <!-- Form Description -->
-            <div class="mb-8 text-sm text-gray-500">
-                Create a DevOps Pipeline for linking Node-RED Instances together.
-            </div>
-
-            <!-- Stage Name -->
-            <FormRow
-                v-model="input.name"
-                type="text"
-                data-form="stage-name"
-            >
-                <template #default>
-                    Stage name
-                </template>
-            </FormRow>
-
-            <!-- Instance -->
-            <FormRow
-                v-model="input.instance"
-                :options="instanceOptions"
-                data-form="stage-instance"
-                :placeholder="instanceDropdownPlaceholder"
-                :disabled="instanceDropdownDisabled"
-            >
-                <template #default>
-                    Choose Instance
-                </template>
-            </FormRow>
-
-            <div class="flex flex-wrap gap-1 items-center">
-                <ff-button
-                    class="ff-btn--secondary"
-                    @click="$router.back()"
-                >
-                    Cancel
-                </ff-button>
-
-                <ff-button
-                    :disabled="!submitEnabled"
-                    :data-action="'create-pipeline'"
-                    type="submit"
-                >
-                    Add Stage
-                </ff-button>
-            </div>
-        </form>
     </main>
 </template>
 
 <script>
 import { ChevronLeftIcon } from '@heroicons/vue/solid'
 
-import ApplicationAPI from '../../../api/application.js'
 import PipelinesAPI from '../../../api/pipeline.js'
 
-import FormRow from '../../../components/FormRow.vue'
 import NavItem from '../../../components/NavItem.vue'
-import SectionTopMenu from '../../../components/SectionTopMenu.vue'
 import SideNavigation from '../../../components/SideNavigation.vue'
 import Alerts from '../../../services/alerts.js'
 
+import PipelineStageForm from './form.vue'
+
 export default {
-    name: 'CreatePipeline',
+    name: 'CreatePipelineStage',
     components: {
         SideNavigation,
-        SectionTopMenu,
         NavItem,
-        FormRow
+        PipelineStageForm
     },
+    inheritAttrs: false,
     props: {
         application: {
             type: Object,
@@ -112,87 +58,30 @@ export default {
             icons: {
                 chevronLeft: ChevronLeftIcon
             },
-            mounted: false,
-            loading: false,
-            instances: [],
-            input: {
-                name: null,
-                instance: null
-            }
+            mounted: false
         }
-    },
-    computed: {
-        submitEnabled () {
-            return this.input.instance && this.input.name
-        },
-        instancesNotInUse () {
-            const instanceIdsInUse = this.pipeline.stages.reduce((acc, stage) => {
-                stage.instances.forEach((instance) => {
-                    acc.add(instance.id)
-                })
-
-                return acc
-            }, new Set())
-
-            return this.instances.filter((instance) => {
-                return !instanceIdsInUse.has(instance.id)
-            })
-        },
-        instanceOptions () {
-            return this.instancesNotInUse.map((instance) => {
-                return {
-                    label: instance.name,
-                    value: instance.id
-                }
-            })
-        },
-        instanceDropdownDisabled () {
-            return this.instancesNotInUse.length === 0
-        },
-        instanceDropdownPlaceholder () {
-            if (this.instancesNotInUse.length === 0) {
-                return 'No instances available'
-            }
-
-            return 'Choose Instance'
-        }
-    },
-    watch: {
-        'application.id': 'loadInstances'
     },
     async mounted () {
         this.mounted = true
-        this.load()
     },
     methods: {
-        async create () {
-            this.loading = true
+        async create (input) {
             const options = {
-                name: this.input.name,
-                instance: this.input.instance
+                name: input.name,
+                instance: input.instance
             }
             if (this.$route.query.sourceStage) {
                 options.source = this.$route.query.sourceStage
             }
             await PipelinesAPI.addPipelineStage(this.$route.params.pipelineId, options)
             Alerts.emit('Pipeline stage successfully added.', 'confirmation')
-            this.loading = false
+
             this.$router.push({
                 name: 'ApplicationPipelines',
                 params: {
                     id: this.application.id
                 }
             })
-        },
-        load () {
-            this.loadInstances()
-        },
-        async loadInstances  () {
-            if (!this.application.id) {
-                return
-            }
-
-            this.instances = await ApplicationAPI.getApplicationInstances(this.application.id)
         }
     }
 }
